requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
backoff==2.2.1
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-debug-toolbar==4.1.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
flake8==6.0.0
googlemaps==4.10.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy==1.4.0
mypy-extensions==1.0.0
numpy==1.25.0
opencage==2.2.0
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
rich==13.4.2
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
uuid==1.30
vine==5.0.0
vk==3.0
wcwidth==0.2.6
